requirements.txt
# Core Framework
streamlit==1.32.0
authlib==1.3.0

# LangChain & RAG
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53

# Embeddings & Vector Store
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# LLM API
groq==0.4.2

# LLM & NLP
transformers==4.35.2
torch==2.1.0
huggingface-hub==0.19.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scipy==1.11.4